apiContainer/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyJWT==2.10.1
requests==2.31.0
